Support cancellation, break apart files, remove hidden internals
