--- dotnet-test-samples/apigw-lambda-list-s3-buckets/src/ServerlessTestSamples.Core/InternalsVisibleTo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly:InternalsVisibleTo("ServerlessTestSamples.UnitTest")]
