+++ dotnet-test-samples/apigw-lambda-list-s3-buckets/src/ServerlessTestSamples.Core/Queries/ListStorageAreasQueryResult.cs
@@ -1,0 +1,8 @@
+﻿namespace ServerlessTestSamples.Core.Queries;
+
+public class ListStorageAreasQueryResult
+{
+    public ListStorageAreasQueryResult(IEnumerable<string> storageAreas) => StorageAreas = storageAreas;
+
+    public IEnumerable<string> StorageAreas { get; }
+}
