+++ dotnet-test-samples/apigw-lambda-list-s3-buckets/src/ServerlessTestSamples.Core/Services/IStorageService.cs
@@ -2,5 +2,5 @@ namespace ServerlessTestSamples.Core.Services;
 
 public interface IStorageService
 {
-    Task<ListStorageAreasResult> ListStorageAreas(string? filterPrefix);
+    Task<ListStorageAreasResult> ListStorageAreas(string? filterPrefix, CancellationToken cancellationToken);
 }
